[Fix] canUseDummyPseudoMain makes too many assumptions
Summary:
It assumed there's a pseudo-main. There isn't for files that fail to parse. It
also assumed that pseudo-mains aren't volatile, which is normally true, but not
if the AllVolatile flag is on. It's on for the Phabricator build, which was
breaking. But pseudo-mains actually shouldn't be volatile, regardless of
AllVolatile.

Test Plan:
slow_tests. Invoke hphp on a file with a parse error, make sure it
doesn't crash. Invoke hphp with AllVolatile on, make sure it doesn't hit the
assert. Build www.

Reviewed By: mwilliams
Reviewers: mwilliams, myang, qigao
Commenters: jungejason
CC: hphp-diffs@lists, jungejason, ps, mwilliams
Revert Plan:
Tags:

- begin *PUBLIC* platform impact section -
Bugzilla: #
- end platform impact -

Differential Revision: 311410