[Fix/HPHPi] Fix bug in ArgArray
Summary:
Assertion was triggered on a number of functions that ArgArray did not
implement. I did another pass and found the cases where we do read-only
operations on ArgArray. New test added to catch where the assertions
failed. There a still a number of assertions remain. They are either
try to mutate the array (which does not make sense for ArgArray and
the ArgArray should have been escalated if such a case), or represents
a misuse of ArgArray (nonSmartCopy, onSetStatic, onSetEvalScalar, and
ArgArray copy constructor).

Test Plan:
make fast_tests
make slow_tests

Reviewers: mwilliams, qigao

Reviewed By: mwilliams

CC: hphp-diffs@lists, ps, mwilliams

Differential Revision: 312913