[Perf/HPHPi] Make static strings unique in HPHPi
Summary:
HPHPi sometimes need to scan the m_alist to find the Variant of a given
name. I tried to make m_alist an Array so the lookup can be faster. However,
it actually hurts performance because of the Array overhead. This is probably
because that most functions do not have a large number of local variables/
parameters to warrant the use of an Array for faster lookup. So instead, I
made a change to ensure that all the StaticStrings are unique. If we have
this property, then we can avoid the cost of doing length check or memcpy
when both names are static (which is highly likely). We may have other
places where a unique StaticString may help to speedup lookups.

Test Plan:
make fast_tests
make slow_tests
flib/unit test
bench.php gets a little under 40s (e.g, 39.791)

Reviewers: mwilliams, qigao

Reviewed By: qigao

CC: hphp-diffs@lists, ps, mwilliams, qigao

Differential Revision: 313940