[Fix] Line number reporting is broken for certain error conditions
Summary:
If a backtrace is taken between the time a new FrameInjection frame is pushed
and its line number is set (e.g. calling an abstract method), the top frame has
a line number of 0, which is useless in error reporting. Omit the frame if it
hasn't had a line number set yet.

Test Plan: fast_tests, slow_tests, test case from the task

Reviewers: mwilliams, myang, qigao

Reviewed By: mwilliams

CC: hphp-diffs@lists, ps, mwilliams

Differential Revision: 312983

Task ID: 689283