[Perf/HPHPi] Create variables on demand
Summary:
This is based on a diff that Stephen has left over and I made some edits.
Currently when calling a function, we prepare a FuncScopeVariableEnvironment.
As part of that, we preallocate all the variables as if they were all needed.
However, if a function has a large number of variables but for a given
invocation only a few of the variables are used, then we would have wasted
time to allocate/deallocate those unused. Secondly, get_defined_vars()
currently shows the variables in their declaration order which is different
from PHP. This change basically does a on-demand allocation of function scope
variables. This is more consistent with PHP and has perf benefit in the case
mentioned above.

Test Plan:
make fast_tests
make slow_tests
flib/unit test
This change has little impact on bench.php but I found that are real cases
this would be more benefitial.

Reviewers: mwilliams, qigao

Reviewed By: mwilliams

CC: hphp-diffs@lists, ps, mwilliams, myang

Differential Revision: 316536