[Fix] Search for constants in the correct order
Summary:
When looking for class constants, hphp would search each base class in turn.
Because of the structure of the code, if the class implemented multiple
interfaces, and neither the parent class, nor the first interface declared the
constant, we would record the constant as missing, even though we went on to
find the constant correctly.

Rewrote to have a recursive step that finds the constant, without reporting
errors, and a (more-or-less) non-recursive step to do the checking.

Test Plan: fast_tests slow_tests, no more false positives for
UseUndeclaredConstant in classes with complex interface hierarchies.

Reviewers: qigao, myang

Reviewed By: myang

CC: ps, mwilliams, myang

Differential Revision: 313830

Task ID: 614967