[Nemo] Dont assume anything about unknown functions
Summary:
The compiler assumes it sees the entire program. This can result in false
assumptions - eg that unknown functions dont mutate their arguments. When
running nemo in diff-at-a-time mode, we need to break that assumption due to
cases like:

  function fiz(array $a) {}
  function bar() {
    $x = "hello";
    foo($x);
    fiz($x);
  }

With the whole program assumption, we assume foo takes its argument by value,
and that fiz is passed a string, rather than an array, and we spit out a warning
about it. But foo may in reality assign an array to $x.

I've added Option::WholeProgram (defaults to true) to fix this. We will
probably want to make more use of this option going forwards.

Test Plan: fast_tests slow_tests, no change to generated files for www-dev.

Reviewers: qigao, myang

Reviewed By: qigao

CC: ps, mwilliams, qigao

Differential Revision: 315584