[perf] objectdata::o_properties should be an array
summary:
for some reason it was an array*, meaning that every use was indirected (and
checked), in addition to the smart pointer indirect and check.

this should be entirely good, since sizeof(array) == sizeof(array*) - so we're
simply eliminating a level of indirection and checking, with no downside.

test plan: fast_tests slow_tests

reviewers: myang, qigao

reviewed by: myang

cc: ps, mwilliams, myang

differential revision: 317713