[server] allow 403 being treated as 404
summary:
as request in task, we want the ability to handle 403 in a way that
people are not sure whether the file exists or not. this diff refactors
the 403 handling code and adds an option to treat 403 as 404.

test plan:
make fast_tests
hit the endpoint from the task, show forbidden when the config is not
set, and show the 404 redirection page when config is set

reviewers: mwilliams, myang

reviewed by: mwilliams

cc: ps, mwilliams

differential revision: 318674

task id: 691676