[fix] fix assertion failure
summary:
<?php
if (true) {
class a {
private $a = 1;
}
class b extends a {
public $a;
function f() { $this->a = 2; }
}
} else {
class a {
protected $a = 1;
}
class b extends a {
public $a;
function f() { $this->a = 2; }
}
}
$obj = new b;
$obj->f();
var_dump($obj);
compile and run with debug=1, assertion `!exists(k)' failed.
the assertion failed because the property name already exists. this can happen
for an overriden property, where the name is first set to null_variant to
occupy the slot for correct iteration order, then set again to the right value.
fixed by using lvalat which does not assume the key is new. new test added.

test plan:
make fast_tests
make slow_tests

reviewers: mwilliams, qigao

reviewed by: mwilliams

cc: hphp-diffs@lists, ps, mwilliams

differential revision: 319212

task id: 701501