[hphpd] fix the backtrace command (w) on large stack
summary:
it used to serialize the whole stack including variables and then omit
the middle, which could take really long time and finally trigger the
exception of hitting stringbuffer limit.
this diff adds a serializewithlimit in variableserializer and makes
debuggerclient call that. this way the backtrace command in stack with
huge variables will not crash the client.
also removed one unused property in the variableserializer class

test plan:
make fast_tests
manually verified it doesn't crash the client in a case which did.

reviewers: myang, mwilliams

reviewed by: myang

cc: ps, mwilliams, myang

differential revision: 320718

task id: 704602