Fix infinite recursion in testcodeerror::testinvalidderivation()
summary:
we were using a case-sensitive hash set to store the set of seen base
classes in classscope::checkderivation(). this was causing infiinte
recursion in test cases for invalid derivation such as the following:

class a extends class a {}

this was because the class name was lower case, whereas the parent name
was upper case. fixed by using a case-insensitive hash map.

test plan: make fast_tests

reviewers: mwilliams, myang

reviewed by: myang

cc: ps, mwilliams, myang

differential revision: 322826

revert plan: ok