[fix/hphpi] fix bug when use $this in closure
summary:
i was too aggressive when i removed the line that enters the current object
into the assoc list as i thought it is already avaiable in the current env.
this fails when $this appears in the use clause of a closure. restored the
line. new test added.

test plan: make fast_tests

reviewers: mwilliams

reviewed by: mwilliams

cc: hphp-diffs@lists, ps, mwilliams, myang

differential revision: 324948