Fix testyield to not access an inaccessible property
summary:
one of the tests in testyield was accessing a protected property 'message'
from the exception class from a context that does not have access. hphpc and
hphpi currently do not do accessibility checks for builtin classes, but we
should still fix the test.

this diff fixes the test to call getmessage() instead of accessing the
'message' property directly.

test plan:
testcoderun testyield
testcoderuneval testyield

reviewers: mwilliams, myang, qigao, kma

reviewed by: mwilliams

cc: ps, mwilliams, andrewparoski

differential revision: 325297