[fix] hphpd crash when throwing non-obj or obj without __tostring()
summary:
hphpd will hit an exception in the catch block when it handles the case
where an non-object is thrown or an object without __tostring method,
causing the debuggerproxy to disconnect. this diff fixes the problem by
adding another catch.

test plan:
make fast_tests
manually run the test in the task and run "throw null"

reviewers: mwilliams

reviewed by: mwilliams

cc: ps, mwilliams

differential revision: 326091

task id: 709819