[fix] fix crash in xbox
summary:
when using xbox for unit test, there was a crash that caused by signal
processing. if the handlers array is already reset and a signal comes,
in the signal handler we may put another arraydata * into the array, which
will no longer be reset before the memory is rolled back. i made a change
to block all signals for this thread so that memory can be safely rolled
back.

test plan:
make fast_tests
make slow_tests
run the unit test using xbox that had the crash, no longer see the crash

reviewers: mwilliams, andrewparoski

reviewed by: mwilliams

cc: hphp-diffs@lists, ps, mwilliams, andrewparoski, myang

differential revision: 327279