Fix testcoderunvm-simplexml
summary:
simplexmlelementiterator holds a raw c++ pointer to its parent, a
simplexmlelement. in expressions like this one:
foreach ($el->children() as $child) { ... }
all references to the simplexmlelement returned by children() will be
dropped after the iterator is created and the object will be
destroyed, leaving the iterator with a stale pointer. it looks like
the iterator already had a smart pointer named m_temp that it stuck its
parent in sometimes; this diff gets rid of m_temp and makes sure the
parent is always in a smart pointer. reset_iterator was split into
set_parent and reset_iterator to avoid unnecessary inc/decrefs in
rewind.

test plan: testsimplexml passes now.

reviewers: kma, andrewparoski, mwilliams

reviewed by: andrewparoski

cc: ps, mwilliams, bsimmers, andrewparoski

differential revision: 326479

task id: 712299