[perf/cleanup/compile-time] miscellaneous code reorg
summary:
various cleanups
 - switch from __gnu_cxx::hash* to std::tr1::unordered*. the former was 72
bytes, and always allocated at least 53 buckets, while the latter is 48 bytes
and starts with 2 buckets (when asked to start with zero). since many of these
maps ended up holding a single class, or a couple of functions this was
extremely wasteful of memory. perf measurements are indicating that its faster
too.
 - change functioncontainer to hold a stringtofunctionscopeptrmap, rather than
*vecmap, since redeclared functions are rare (non-existent in classes, and
extremely unusual in file-scope). hold a separate stringtofunctionscopeptrvecmap
(allocated on demand) to deal with the case of filescope redeclared functions.
 - get rid of .fw.h headers. they are only ever included from one place (the
corresponding .h) and are often empty. this reduces the number of generated
files by about 40%.
 - remove various unneeded forward declarations, and move some virtual
functions out of headers into the .cpp files.

test plan: fast_tests slow_tests

reviewers: qigao, myang

reviewed by: myang

cc: achao, ps, mwilliams, myang

differential revision: 326872