[perf/hphpi] reduce number of user function call lookup
summary:
when calling a user function in hphpi through simplefunctioncall,
even though we know the name at compile time, we still do a lookup
to find the function statement at runtime. if a function is called
in a loop, the cost of doing lookup becomes more significant. i added
a function table that can be used to reduce lookups. each user
function name is assigned a unique id (can be reused in case reparsing)
that is used as the index into the function table which is a vector.
whenever a function declared, we also enter the functionstatement into
the function table. in simplefunctioncall, we cache the id for the
function being called and use it as a faster way to call the function
if the function is already declared. to prevent a runaway id, i put
a cap as a runtime option. after hitting the cap, no id will be assigned
and those functions have to be called via normal name lookup.

test plan:
make fast_tests
make slow_tests
flib/unit test
micro benchmark:
<?php
$start = microtime(true);
function foo() { }
for ($i = 0; $i < 10000000; $i++) foo();
$end = microtime(true);
echo $end - $start . "\n";
before:
2.4777600765228
after
2.0086898803711

reviewers: mwilliams, qigao

reviewed by: mwilliams

cc: hphp-diffs@lists, ps, mwilliams, myang

differential revision: 330013