[perf] better class constants
summary:
if a class constant was defined in terms of something that was not itself a
scalar (eg another class constant), we immediately marked it dynamic, even
though the initializer might eventually be resolved to a scalar.

during optimization, we would only replace a class constant if the value was
scalar, when we can get better results by always replacing.

type inference allways assumed variant for dynamic constants, but thats not
necessary. eg "class x { const foo=1; } class y { const bar=x::foo; }". if x is
volatile, then y::bar is dynamic. but it is known to be of type int, and to have
the value 1. its only dynamic because it might fatal when accessed (if x doesnt
exist).

any class which contained at least one dynamic constant was marked as
lazy-init, which affected every use of every constant/static in the class.

test plan: fast_tests slow_tests

reviewers: qigao, myang

reviewed by: qigao

cc: ps, mwilliams, qigao

differential revision: 336136