[fix/perf/hphpi] add objectstringpropertyexpression
summary:
almost all objectpropertyexpression has a explicit string as the property
name. i added objectstringpropertyexpression to capture the remaining
cases where the object may not be $this or a simple variable expression
but the property is a string. i also fixed a bug where m_reverseorder
wasn't set (caused by one of my earlier changes) although i couldn't
find a case where this matters because the use of tempexpressionlist
when eval order matters so it may not be needed but leaving it uninited
does not seem right.

test plan:
make fast_tests
bench.php showed about 0.5s speed up.

reviewers: mwilliams, qigao

reviewed by: mwilliams

cc: hphp-diffs@lists, ps, mwilliams

differential revision: 337080