[fix] turn off argarray
summary:
the design of argarray was flawed because it does not consider the memory
rollback and checkpoint mechanism properly and can leave the memory in
inconsistent state. i am turning it off for now and will reevaluate whether
it is worthwhile to leave it because the removal of collect_few_args_ref
has made it a rare case of packing arguments into an array.

test plan:
make fast_tests
run hphpi server mode test several times with no crashes

reviewers: mwilliams, qigao

reviewed by: mwilliams

cc: hphp-diffs@lists, ps, mwilliams, myang

differential revision: 339617