[fix] folding concats needs to preserve conversion to string
summary:
the code to fold concats was checking the wrong condition (at one point, this
condition would have worked, but it assumed that nothing touched the expected
type on the node being replaced - this was too fragile, and is no longer true).
instead, we check that the expected type on the concat is string - and if not we
insert a string cast.

there was a similar issue with $x + 0 and $x * 1, but those were sufficiently
conservative that i couldnt actually reproduce a problem. the new checks should
be both correct, and less conservative.

test plan: fast_tests slow_tests

reviewers: qigao, myang, andrewparoski

reviewed by: myang

cc: ps, mwilliams, myang

differential revision: 340359

task id: 752902