Add thunk logic to the hphpi binary
summary:
this diff adds thunk logic so that we can make the hphpi binary redirect
to another binary via execv(). the idea here is that we can set a symlink
"/.hphpi-thunk" on individual machines to make hphpi redirect to a different
binary.

note that the thunk will only be triggered for the hphpi binary, and only
if "/.hphpi-thunk" exists in the filesystem. other binaries produced by
hphpc will not trigger the thunk and will continue to behave the same as
before.

blame rev:

reviewers: je, kma, mwilliams, qigao, oyamauchi

cc:

test plan:
fast_tests
slow_tests
verify the thunk is triggered for the hphpi binary
vierfy the thunk is not triggered for other binaries
build www and browse the site

revert plan:

differential revision: 340844