[fix] allow private static properties to be found by dynamic lookup
summary:
os_lval and os_get have always excluded private properties, presumably because
they would normally be resolved statically. there are cases however (eg
involving redeclared classes, or $cls::$property) where we cant resolve it
statically, and so we should allow privates to be found.

this fix isnt exactly right, since it could find a public property in a more
derived class, when it should have found a private property with the same name
(it also allows access to a private property that should be hidden). but this
fixes the immediate issue.

test plan: fast_tests slow_tests

reviewers: kma, qigao, myang

reviewed by: myang

cc: ps, mwilliams, myang

differential revision: 342147