[fix] only optimize dynamicnamedcall when its known to be dynamic
summary: there was a small optimization for the case of a dynamic call through
a derived-from-redeclared class. but its only valid when the call is known to be
dynamic. the objcall flag is used to indicate dynamic (>0) possibly-dynamic
(<0), and static (0), but we were just testing for non-zero. replace with a test
for >0 instead.

test plan: fast_tests slow_tests

reviewers: myang, kma

reviewed by: myang

cc: ps, mwilliams, myang

differential revision: 342326