Debugger server api
summary:
to build gui for hphpd, we need to access debuggerclient using an api.
this diff changes the existing debugger extension to do that. the
current built-in class debuggerclient and debuggerproxy are just to
support cmduser, so they are renamed with suffix cmduser. then a new
extension class is added to wrap around debuggerclient
for executing commands. an extension function is added to fetch
debuggerclient by name so that multiple requests can operate on the same
client for debugging. this way, a gui tool can use ajax request to hit
the server and operate on debugger via this api.

since this change makes debuggerclient running in server mode as root,
it breaks a few assumptions in debuggerclient, so several other
modifications are added to make sure:
- the sweeping after request doesn't break anything,
- multiple debuggerclient can co-exist on the same server without
interfering each other
- credentials do not rely on local login and home
- disable color and highlighting since it's not terminal printing

this diff also modifies the makefile for idl the same way as makfile for
system so that it doesn't give a lot of warnings.

some more work need to be done for better handling error and corner
cases. also need to find a way to format output. for now it just
returns the string printed by debuggerclient after each processcmd().

test plan:
make fast_tests
tried most functionality with a debug php endpoint on sandbox

reviewers: myang, mwilliams

reviewed by: myang

cc: tuomaspelkonen, ps, mwilliams, myang

differential revision: 346215

task id: 753544