[perf] reduce temporary generation
summary:
fix a few issues:
 - if any child node had a side effect, and produces a temporary anyway, we
would enforce ordering. but this isnt needed for a single child node, and is
counter-productive for eg isset and empty.
 - if we wanted to output a fast-cast, we would force a temporary. but this is
wrong in exist context for arrayelementexpression and objectproperty expression
(the temporary will be evaluated using rvalat, rather than isset/empty). so in
these cases, we would emit both an rvalat, and an isset/empty.
 - type inference was wrong for static method calls to non-static functions,
when there is (or might be) no object available (it failed to set
implementedtype to variant, which it needs to because the call will be
dispatched via an invoke helper).

test plan: fast_tests slow_tests

reviewers: myang, qigao

reviewed by: myang

cc: ps, mwilliams, myang

differential revision: 345684