[fix] do not set http_raw_post_data if it cause string overflow
summary:
in video upload, the amount of data can overflow a string and cause exception
to be throw. i made a change so that we don't store http_raw_post_data when
this happens.

test plan:
make fast_tests
upload a large video in non-chunk mode and make sure no exception is thrown
at the end.

reviewers: mwilliams, qigao

reviewed by: mwilliams

cc: hphp-diffs@lists, ps, mwilliams

differential revision: 345263