[perf] reduce o_set overhead
summary:
objectdata's o_set family of functions all took a bool flag "forinit". the flag
was used to bypass access checks, and was only used when initializing an
objectdata from hphpi, and from o_setarray. this meant that we were always
passing in, and checking a mostly unused parameter. in addition, object's o_set
family didnt have the flag. code-gen didnt differentiate the two cases, and so
in some cases was passing the class name as the "forinit" flag. bizarrely, this
mostly worked, because the class-name would be converted to true, and access
would be allowed (which is usually the right thing to do in correct programs).

with this diff, objectdata's o_set family matches the object o_set family,
solving the correctness problem, and also reducing the call overhead by dropping
a parameter.

test plan: fast_tests slow_tests

reviewers: qigao, myang

reviewed by: myang

cc: ps, mwilliams, myang, qigao

differential revision: 346544