[fix] misc reference handling buglets exposed by enabling inlining
summary:
 - in f($x, $x &= $y), where f takes its first param by reference, we would
   pass a reference to $x after the assignment, rather than before it (this
   happened independent of inlining).
 - in f(g(x)) if g returned by reference, and f did not take its param by
   reference, f could end up with a reference to the result of g anyway, due to
   copy constructor elision (again, independent of inlining)
 - there was a similar problem when a return-by-reference function got inlined.
   the return value could be destroyed before it was used (resulting in a crash in
   one of the added test cases), or it could just end up refering to the wrong
   thing.
 - in $a = &f(), if f returned by value, and got inlined, $a could end up bound
   to the result of f() (rather than bound to a copy of the result of f())

in order to expose the return-by-reference issue, i had to increase the
inliner's agressiveness. so ive changed it from a bool, to an int and used it as
the max "cost" for a function.

to minimize extra temps, i added the return value's refness to the functioninfo.

test plan: fast_tests, slow_tests and slow_tests with autoinline=5 (there are
still some failures, with the latter - will address separately).

reviewers: qigao, myang

reviewed by: myang

cc: ps, mwilliams, myang

differential revision: 348039

task id: 630786