[fix/hphpd] fix bugs in server api
summary:
this diff fixes the issue of interruping for the second time causing the
server to abort. it also makes the state machine more clear and guards
mroe strict so that user retries are unlikely to cause issues and leave
server in an error state.

note: there are some generated files change that are not because of this
diff.

test plan:
make fast_tests
manually tried more functionalities

reviewers: myang

reviewed by: myang

cc: tuomaspelkonen, ps, mwilliams, myang

differential revision: 350417

task id: 753544