[fix/hphpd] remove the trailing ';' for '=' command
summary:
in a recent change, () were added to wrap the expression for '=', which
made ';' in the expression a syntax error. we were thinking on a general
fix but it seems hard to do without calling php parser for the command
line. this diff fixes one common benign case with traling ';', which used
to work before the () were added.

test plan:
make fast_tests
tried with hphpd on command line
hphpd> =123;
123

reviewers: mwilliams, myang

reviewed by: myang

cc: ps, mwilliams, myang, qigao

differential revision: 351726

task id: 747161