[fix] guard analysisresult::m_files with a lock
summary: for parse-on-demand, we were checking to see if a file had already
been parsed without aquiring a lock. that could result in a crash if files were
inserted during the lookup.

test plan: running the command from the task resulted in a crash once in about
tries before the fix. after the fix i ran it 3000 times with no failures.

reviewers: qigao, myang

reviewed by: qigao

cc: ps, mwilliams, qigao

differential revision: 351734

task id: 778331