[memcachepool extension] changing mode of closing connections to not close tcp connection on every memcache petition, but only at the end on the requests.
this turns a little hard to do because the particular behaviour of request init
and shutdown functions of hiphop, which requires you to access some variable of
the requesteventhandler-derived class in order to register it on the request.

in this case, either you access some method of memcachepoolrequest or the
requestshutdown is not called at the end on the request.

see src/runtime/base/util/request_local.h