Fix eof detection on socket connections
A flag was being used to mark the Socket object as eof when reading, but
this have some behaviour differences with PHP. On PHP you don't need to
read from a socket in order to detect an eof situation. We are using here
the trick of doing a recv with MSG_PEEK to detect the situation.