Make the takeover transition smoother
During the takeover proccess, sometimes the leaving hiphop servers generates a
core due the double stop of the admin server (because they receive a /stop order
while the admin server is stopping).

We changed the steps a bit to be more gentle and wait a bit for the old server
to just stop themself. Also, another extra protection is added to avoid the
segfault in case of a double stop happens again.